requirements.txt
Django>=2.0
graphene-django==2.0.0
psycopg2-binary
Redis
